requirements.txt

fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.1
google-genai>=0.3.0
